correcao estrutura layouts interno e externo
