--- src/app/components/template/footer/footer.component.css
@@ -1,16 +1,0 @@
-.footer {
-    position: fixed;
-    bottom: 0;    
-
-    display: flex;
-    justify-content: flex-end;
-}
-
-.footer > span {
-    font-size: 1rem;
-    font-weight: 300;
-}
-
-.red {
-    color: red;
-}
