--- src/app/components/template/footer/footer.component.html
@@ -1,5 +1,0 @@
-<mat-toolbar class="footer">
-    <span>
-       Todos os direitos reservados
-    </span>       
-</mat-toolbar>
