--- src/app/components/template/nav/nav.component.html
@@ -1,32 +1,0 @@
-<mat-sidenav-container class="container">
-    <mat-sidenav class="sidenav" mode="side" opened 
-        fixedInViewport="true" fixedTopGap="64">
-        <mat-nav-list class="nav-list">
-            <a mat-list-item routerLink="/">
-                <i class="material-icons">
-                    home
-                </i>
-                Home
-            </a>            
-            <a mat-list-item  routerLink="/product">
-                <i class="material-icons">
-                    storefront
-                </i>
-                Produtos
-            </a>
-            <a mat-list-item routerLink="/">
-                <i class="material-icons">
-                    account_box
-                </i>
-                Usuário                
-            </a>           
-        </mat-nav-list>
-    </mat-sidenav>
-
-
-    <!-- CONTEÚDO-->
-    <mat-sidenav-content class="content">   
-        <router-outlet></router-outlet>
-    </mat-sidenav-content>
-
-</mat-sidenav-container>
