+++ src/app/external/external.component.css
@@ -1,0 +1,3 @@
+.container {
+    padding: 5em;    
+}
