+++ src/app/external/external.component.html
@@ -1,0 +1,3 @@
+<div class="container">
+    <router-outlet></router-outlet>
+</div>
