+++ src/app/external/external.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-external',
+  templateUrl: './external.component.html',
+  styleUrls: ['./external.component.css']
+})
+export class ExternalComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
