+++ src/app/external/external.module.ts
@@ -1,0 +1,58 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { HomeComponent } from './home/home.component';
+import { LoginComponent } from './login/login.component';
+import { RegisterComponent } from './register/register.component';
+import { ExternalComponent } from './external.component';
+import { ExternalRoutingModule } from './external-routing.module';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatTableModule } from '@angular/material/table';
+import { MatIconModule } from '@angular/material/icon';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatSortModule } from '@angular/material/sort';
+
+
+
+@NgModule({
+  declarations: [
+    ExternalComponent,
+    HomeComponent,
+    LoginComponent,
+    RegisterComponent,        
+  ],
+  imports: [
+    CommonModule,
+    ExternalRoutingModule,    
+    FormsModule,
+    ReactiveFormsModule,
+    MatToolbarModule,
+    MatSidenavModule,
+    MatListModule,
+    MatCardModule,
+    MatButtonModule,
+    MatSnackBarModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatTableModule,
+    MatIconModule,
+    MatPaginatorModule,
+    MatSortModule,   
+  ],
+  exports: [
+    ExternalComponent,    
+    HomeComponent,
+    LoginComponent,
+    RegisterComponent,   
+  ]
+})
+export class ExternalModule { }
