+++ src/app/external/home/home.component.html
@@ -1,0 +1,12 @@
+<h1>Bem vindo!</h1>
+
+<h4> Sistema de exemplo CRUD com Angular.</h4>
+
+<a [routerLink]="['/login']" routerLinkActive="router-link-active" mat-raised-button class="button-add"
+    style="margin-right: 2em;">
+    Login
+</a>
+
+<a [routerLink]="['/register']" routerLinkActive="router-link-active" mat-raised-button>
+    Register
+</a>
