+++ src/app/external/register/register.component.html
@@ -1,0 +1,47 @@
+<mat-card>
+    <mat-card-title>Register</mat-card-title>
+
+    <form [formGroup]="form" (ngSubmit)="onSubmit()">
+
+        <mat-form-field>
+            <mat-label>Username</mat-label>
+            <input matInput formControlName="username">
+            <mat-error *ngIf="submitted && frm.username.invalid">
+                <p class="alert">Nome deve conter no mínimo 4 caracteres!</p>
+            </mat-error>
+        </mat-form-field>
+        
+        <mat-form-field>
+            <mat-label>Password</mat-label>
+            <input matInput formControlName="password">
+            <mat-error *ngIf="submitted && frm.password.invalid">
+                <p class="alert">Informe a senha!</p>
+            </mat-error>
+        </mat-form-field>
+
+        <mat-form-field>
+            <mat-label>Confirm Password</mat-label>
+            <input matInput formControlName="confirmPassword">
+            <mat-error *ngIf="submitted && frm.confirmPassword.invalid">
+                <p class="alert">Informe a confirmação da senha!</p>
+            </mat-error>
+        </mat-form-field>
+
+        <div class="buttons">
+            <button type="submit" mat-raised-button class="button-add">
+                Register
+            </button>
+
+            <a mat-raised-button routerLink="/login">
+                Login
+            </a>
+        </div>
+        
+        <div class="buttons">
+            <a [routerLink]="['/home']" routerLinkActive="router-link-active">
+                Fechar
+            </a>
+        </div>
+
+    </form>
+</mat-card>
