+++ src/app/internal/internal.component.html
@@ -1,0 +1,2 @@
+<app-header></app-header>
+<app-nav></app-nav>
