+++ src/app/internal/product/product-list2/product-list2.component.html
