+++ src/app/internal/product/product-list2/product-list2.component.ts
@@ -1,0 +1,39 @@
+import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatTable } from '@angular/material/table';
+import { Router } from '@angular/router';
+import { Product } from 'src/app/models/product';
+import { ProductService } from 'src/app/services/product.service';
+import { ProductDataSource } from './product-list2-datasource';
+
+@Component({
+  selector: 'app-product-list2',
+  templateUrl: './product-list2.component.html',
+  styleUrls: ['./product-list2.component.css', '../product.component.css']
+})
+export class ProductList2Component implements AfterViewInit, OnInit {
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatTable) table: MatTable<Product>;
+  dataSource: ProductDataSource;
+
+  /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
+  displayedColumns = ['name', 'price', 'action']
+
+  constructor(private productService: ProductService, private router: Router) { }
+
+  ngOnInit() {
+    this.dataSource = new ProductDataSource(this.productService);
+  }
+
+  create(): void {
+    this.router.navigate(['/product/create'])
+  }  
+
+  ngAfterViewInit() {
+    this.dataSource.sort = this.sort;
+    this.dataSource.paginator = this.paginator;
+    this.table.dataSource = this.dataSource;
+  }
+}
