+++ src/app/internal/product/product.component.ts
