+++ src/app/internal/users/users.component.html
@@ -1,0 +1,1 @@
+<p>users works!</p>
