--- src/app/pages/home/home.component.html
@@ -1,4 +1,0 @@
-
-<h1>Bem vindo!</h1>
-
-<h4> Sistema de exemplo CRUD com Angular.</h4>
