+++ src/app/template/nav/nav.component.html
@@ -1,0 +1,32 @@
+<mat-sidenav-container class="container">
+    <mat-sidenav class="sidenav" mode="side" opened 
+        fixedInViewport="true" fixedTopGap="64">
+        <mat-nav-list class="nav-list">
+            <a mat-list-item routerLink="/dashboard">
+                <i class="material-icons">
+                    dashboard
+                </i>
+                Dashboard
+            </a>            
+            <a mat-list-item  routerLink="/product">
+                <i class="material-icons">
+                    storefront
+                </i>
+                Produtos
+            </a>
+            <a mat-list-item routerLink="/users">
+                <i class="material-icons">
+                    supervised_user_circle
+                </i>
+                Usuários               
+            </a>           
+        </mat-nav-list>
+    </mat-sidenav>
+
+
+    <!-- CONTEÚDO-->
+    <mat-sidenav-content class="content">   
+        <router-outlet></router-outlet>
+    </mat-sidenav-content>
+
+</mat-sidenav-container>
